refactor: Move performGenerate to Analyzers and hide access to individual analyzers
