+++ tools/serverpod_cli/lib/src/commands/start.dart
@@ -21,8 +21,8 @@ import 'package:serverpod_cli/src/commands/start/tui/tui_log_sink.dart';
 import 'package:serverpod_cli/src/commands/start/tui/tui_logger.dart';
 import 'package:serverpod_cli/src/commands/start/watch_session.dart';
 import 'package:serverpod_cli/src/commands/watcher.dart';
+import 'package:serverpod_cli/src/generator/analyzers.dart';
 import 'package:serverpod_cli/src/generator/generation_staleness.dart';
-import 'package:serverpod_cli/src/generator/generator.dart';
 import 'package:serverpod_cli/src/generator/isolated_analyzers.dart';
 import 'package:serverpod_cli/src/runner/serverpod_command.dart';
 import 'package:serverpod_cli/src/runner/serverpod_command_runner.dart';
